fix Фича: При пергрузке пользователя задачами выдаст предупреждение

+++ src/main/java/com/site/webapp/controllers/TaskController.java
@@ -89,7 +89,7 @@ public class TaskController extends LoggingController{
 
     @GetMapping("/api/artists/{id}/active-tasks")
     public String checkArtistLoad(@PathVariable Long id, Model model) {
-        int activeTasks = userService.getActiveTaskCount(id);
+        long activeTasks = userService.getActiveTaskCount(id);
         if (activeTasks >= 5) { // ваш лимит
             model.addAttribute("warning", "Внимание: у исполнителя уже " + activeTasks + " активных задач");
         }
